+++ test/e2e/app-websocket.ts
@@ -12,6 +12,7 @@ import {
   RoleName,
   RoleNameCallZomeRequest,
   SignalCb,
+  SignalType,
 } from "../../src";
 import {
   createAppWsAndInstallApp,
@@ -90,7 +91,7 @@ test(
       (resolve) => (resolveSignalPromise = resolve)
     );
     const signalCb: SignalCb = (signal) => {
-      assert(signal.type === "app");
+      assert(signal.type === SignalType.App);
       t.deepEqual(signal.value, {
         cell_id,
         zome_name: TEST_ZOME_NAME,