+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+
+### Added
+### Fixed
+### Changed
+### Removed
+
+## 2025-04-18: v0.19.0-rc.1
 ### Added
 - Tests for magic config variables: `window.__HC_LAUNCHER_ENV__` and `window.__HC_ZOME_CALL_SIGNER__`
 - Added `SignalType` enum